refactor: stripMarkdownをlib/に切り出し
routes/line.tsから汎用ユーティリティとしてlib/strip-markdown.tsに移動した。
